[Refactor] #111 - View Depth 줄이기. 너무 분리된 View들을 다시 합쳐주었음.
