--- WSSiOS/WSSiOS/Source/Presentation/Register/RegisterNormal/RegisterNormalView/RegisterNormalAssistantView/RegisterNormalNovelInfoView.swift
@@ -1,110 +1,0 @@
-//
-//  RegisterNormalNovelInfoStackView.swift
-//  WSSiOS
-//
-//  Created by 이윤학 on 1/10/24.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class RegisterNormalNovelInfoView: UIView {
-    
-    // MARK: - UI Components
-    
-    private let novelInfoStackView = UIStackView()
-    private let novelTitleLabel = UILabel()
-    private let novelAuthorLabel = UILabel()
-    
-    // MARK: - Life Cycle
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        setUI()
-        setHieararchy()
-        setLayout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - Custom Method
-    
-    private func setUI() {
-        novelInfoStackView.do {
-            $0.axis = .vertical
-            $0.spacing = 6
-            $0.alignment = .center
-        }
-        
-        novelTitleLabel.do {
-            // 로딩 전&후 화면의 이질감을 줄이기 위해 로딩중 문구 추가
-            $0.text = StringLiterals.Loading.novelTitle
-            novelTitleAttribute(of: $0)
-        }
-        
-        novelAuthorLabel.do {
-            // 로딩 전&후 화면의 이질감을 줄이기 위해 추가
-            $0.text = StringLiterals.Loading.novelAuthor
-            novelAuthorAttribute(of: $0)
-        }
-    }
-    
-    private func setHieararchy() {
-        self.addSubview(novelInfoStackView)
-        novelInfoStackView.addArrangedSubviews(novelTitleLabel,
-                                               novelAuthorLabel)
-    }
-    
-    private func setLayout() {
-        novelInfoStackView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-    }
-    
-    func bindData(title: String?, author: String?) {
-        novelTitleLabel.do {
-            $0.text = title
-            novelTitleAttribute(of: $0)
-        }
-        
-        novelAuthorLabel.do {
-            $0.text = author
-            novelAuthorAttribute(of: $0)
-        }
-    }
-    
-    /// novelTitle의 UI를 위한 함수.
-    private func novelTitleAttribute(of label: UILabel) {
-        label.do {
-            $0.makeAttribute(with: label.text)?
-                .lineSpacing(spacingPercentage: 140)
-                .kerning(kerningPixel: -1.2)
-                .applyAttribute()
-            $0.font = .HeadLine1
-            $0.textColor = .White
-            $0.textAlignment = .center
-            $0.lineBreakMode = .byTruncatingTail
-            $0.lineBreakStrategy = .hangulWordPriority
-            $0.numberOfLines = 3
-        }
-    }
-    
-    /// novel의 AttributedText 적용을 위한 함수.
-    private func novelAuthorAttribute(of label: UILabel) {
-        label.do{
-            $0.makeAttribute(with: label.text)?
-                .lineSpacing(spacingPercentage: 150)
-                .kerning(kerningPixel: -0.6)
-                .applyAttribute()
-            $0.font = .Body2
-            $0.textColor = .Gray200
-            $0.numberOfLines = 1
-            $0.lineBreakMode = .byTruncatingTail
-        }
-    }
-}
