+++ WSSiOS/WSSiOS/Source/Presentation/Register/RegisterNormal/RegisterNormalView/RegisterNormalAssistantView/RegisterNormalNovelSummaryView.swift
@@ -12,14 +12,22 @@ import Then
 
 final class RegisterNormalNovelSummaryView: UIView {
     
+    //MARK: - Properties
+
+    var platformList: [UserNovelPlatform] = []
+
     // MARK: - UI Components
     
     private let novelSummaryStackView = UIStackView()
+    
     private let plotTitleLabel = WSSSectionTitleView()
     private let plotLabel = UILabel()
+    
     private let genreTitleLabel = WSSSectionTitleView()
     private let genreLabel = UILabel()
-    let platformView = RegisterNormalPlatformView()
+    
+    private let platformTitleLabel = WSSSectionTitleView()
+    let platformCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewLayout())
     
     // MARK: - Life Cycle
     
@@ -54,6 +62,21 @@ final class RegisterNormalNovelSummaryView: UIView {
         genreTitleLabel.do {
             $0.setText(StringLiterals.Register.Normal.SectionTitle.genre)
         }
+        
+        platformTitleLabel.do {
+            $0.setText(StringLiterals.Register.Normal.SectionTitle.platform)
+        }
+        
+        platformCollectionView.do {
+            let layout = LeftAlignedCollectionViewFlowLayout()
+            layout.scrollDirection = .vertical
+            layout.minimumLineSpacing = 8
+            layout.minimumInteritemSpacing = 8
+
+            $0.collectionViewLayout = layout
+            $0.isScrollEnabled = false
+            $0.backgroundColor = .clear
+        }
     }
     
     private func setHieararchy() {
@@ -62,7 +85,8 @@ final class RegisterNormalNovelSummaryView: UIView {
                                                   plotLabel,
                                                   genreTitleLabel,
                                                   genreLabel,
-                                                  platformView)
+                                                  platformTitleLabel,
+                                                  platformCollectionView)
     }
     
     private func setLayout() {
@@ -75,12 +99,36 @@ final class RegisterNormalNovelSummaryView: UIView {
             
             $0.setCustomSpacing(10, after: plotTitleLabel)
             $0.setCustomSpacing(10, after: genreTitleLabel)
+            $0.setCustomSpacing(10, after: platformTitleLabel)
+        }
+        
+        platformCollectionView.snp.makeConstraints {
+            $0.leading.trailing.equalToSuperview()
+            $0.bottom.equalToSuperview()
+            $0.height.equalTo(100)
         }
     }
     
     func setText(of label: UILabel, text: String?) {
         label.text = text
         bodyStyle(of: label)
+    }
+    
+    func bindData(plot: String?, genre: String?, platforms: [UserNovelPlatform]) {
+        setText(of: plotLabel, text: plot)
+        setText(of: genreLabel, text: genre)
+        platformList = platforms
+        
+        if platformList.count == 0 {
+            platformTitleLabel.isHidden = true
+            platformCollectionView.isHidden = true
+        }
+    }
+    
+    func updateCollectionViewHeight(height: CGFloat) {
+        platformCollectionView.snp.updateConstraints {
+            $0.height.equalTo(height)
+        }
     }
     
     /// 각 Section의 본문 텍스트 스타일
@@ -94,11 +142,5 @@ final class RegisterNormalNovelSummaryView: UIView {
             $0.numberOfLines = 0
             $0.textColor = .Gray300
         }
-    }
-    
-    func bindData(plot: String?, genre: String?, platforms: [UserNovelPlatform]) {
-        setText(of: plotLabel, text: plot)
-        setText(of: genreLabel, text: genre)
-        platformView.bindData(platforms: platforms)
     }
 }
